+++ src/main/java/roomescape/domain/reservation/dto/ReservationTimeAddRequest.java
@@ -1,7 +1,7 @@
 package roomescape.domain.reservation.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.reservation.domain.reservationTim.ReservationTime;
+import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
 
 public record ReservationTimeAddRequest(LocalTime startAt) {
     public ReservationTime toEntity() {
